filter: Remove "decimated lines" accounting
It's clearly lying, because we can round-trip an already-optimized
file and get the same lines out (great), but it will claim we dropped
tens of thousands of lines (not great).

My guess is this has to do with the decimation timekeeping;
"dropping" a line that occurs at prev + rate,
only to immediately write another one when we pass that timestamp.